perf(mge/imperative): misc optimizations
GitOrigin-RevId: bbe7a10b007e8b6d8a66dd64ff83ac1df4b9f8d2

+++ imperative/python/megengine/functional/math.py
@@ -229,7 +229,7 @@ def mean(
         [3.5]
 
     """
-    return inp.astype("float32").mean(axis=axis, keepdims=keepdims)
+    return inp.mean(axis=axis, keepdims=keepdims)
 
 
 def var(
